unlink unless not exists

+++ lib/tasks/backup.rake
@@ -6,7 +6,7 @@ namespace :backup do
   task :config => :environment do
     zipfile_path = Rails.root.join('tmp', 'config_yml.zip')
 
-    File.unlink(zipfile_path)
+    File.unlink(zipfile_path) unless !File.exists?(zipfile_path)
 
     Zip::File.open(zipfile_path, Zip::File::CREATE) do |zipfile|
       Dir[Rails.root.join('config', '*.yml')].each do |f|
